Updated, Generated and Validated AllergyIntolerance JSON files
The following changes were made to the master AllergyIntolerance test data spreadsheet:
1. criticality = "medium" is not a member of the bound (required) AllergyIntoleranceCriticality value set. penicillin2 criticality changed to "high"; aspirin criticality changed to "unable-to-assess".
2. "asprin" typo. Changed to "aspirin".
3. mmr allergy code.text - added "vaccine" to clarify that the allergy is to the "measles/mumps/rubella [MMR] vaccine" rather than the "measles/mumps/rubella [MMR]" which are the diseases.

It was noted (but not changed) that AllergyIntolerance-nkfa has a verificationStatus of "confirmed" which seems to be contrary to protracted TDG discussions about negated allergies/intolerances like nka, nkfa, nkda not having a confirmed status as it's near impossible to test/prove the patient has no allergy to any food or substance. Personally, I didn't agree with the proposed guidance and interpretation of the verificationStatus concepts definitions and the guidance was removed however just double checking that the confirmed verificationStatus was intended for this nkfa resource.

Confirming that the FHIR JSON files were generated from the csv into the "generated" folder and validated using the Java validator with no errors.